requirements.txt
# LLM Integration
# Using OpenAI SDK with OpenRouter.ai for flexible model switching (GPT-4, Claude, etc.)
openai>=1.10.0

# Data Processing
pydantic>=2.5.0
pyyaml>=6.0
python-frontmatter>=1.0.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.12.0
ruff>=0.1.9

# Additional utilities
requests>=2.31.0

# M2: RAG & Vector Search
sentence-transformers==2.2.0  # Local embedding model for CV chunking
numpy>=1.24.0                  # Vector operations for retrieval
scipy>=1.11.0                  # Cosine similarity calculations

# M2: Streamlit UI (Demo & Debugging)
streamlit==1.31.0              # Inspection UI for RAG pipeline visualization
pandas>=2.0.0                  # DataFrame support for Streamlit tables
